- WITH_CXX_GUARDEDALLOC working again - CMake building without python or fluidsim working again (broke in recent commit) - remove BLI_short_filename(), it wasnt used anywhere.

+++ source/gameengine/GameLogic/SCA_LogicManager.h
@@ -146,7 +146,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:SCA_LogicManager"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:SCA_LogicManager"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
