+++ source/gameengine/Ketsji/KX_TimeLogger.h
@@ -106,7 +106,7 @@ protected:
 
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:KX_TimeLogger"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:KX_TimeLogger"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
