+++ source/gameengine/Network/NG_NetworkDeviceInterface.h
@@ -74,7 +74,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:NG_NetworkDeviceInterface"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:NG_NetworkDeviceInterface"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
