+++ source/gameengine/Physics/common/PHY_IMotionState.h
@@ -59,7 +59,7 @@ class	PHY_IMotionState
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:PHY_IMotionState"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:PHY_IMotionState"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
