+++ source/gameengine/Rasterizer/RAS_Rect.h
@@ -96,7 +96,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:RAS_Rect"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:RAS_Rect"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
