+++ source/gameengine/SceneGraph/SG_DList.h
@@ -215,7 +215,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_DList"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_DList"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
