+++ source/gameengine/SceneGraph/SG_Tree.h
@@ -115,7 +115,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_Tree"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_Tree"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
@@ -158,7 +158,7 @@ public:
 	
 #ifdef WITH_CXX_GUARDEDALLOC
 public:
-	void *operator new( unsigned int num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_TreeFactory"); }
+	void *operator new(size_t num_bytes) { return MEM_mallocN(num_bytes, "GE:SG_TreeFactory"); }
 	void operator delete( void *mem ) { MEM_freeN(mem); }
 #endif
 };
